Refactor section properties and add new section beverages
