+++ iOS-GreenRecipes.xcodeproj/xcuserdata/anupamberi.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -370,101 +370,5 @@
             </Locations>
          </BreakpointContent>
       </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "38A54524-2FD0-4903-B776-1F8058C591E3"
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "iOS-GreenRecipes/View Controllers/RecipesHomeViewController/RecipesHomeViewController.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "283"
-            endingLineNumber = "283"
-            landmarkName = "applyInitialSnapshots()"
-            landmarkType = "7">
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "E7EB5D3F-2D00-429F-8AB0-AAC58835FA1B"
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "iOS-GreenRecipes/View Controllers/RecipesHomeViewController/RecipesHomeViewController.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "190"
-            endingLineNumber = "190"
-            landmarkName = "createLayout()"
-            landmarkType = "7">
-            <Locations>
-               <Location
-                  uuid = "E7EB5D3F-2D00-429F-8AB0-AAC58835FA1B - 157e80af01ba7bdd"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "iOS_GreenRecipes.RecipesHomeViewController.createRecipeCellRegistration() -&gt; __C.UICollectionView.CellRegistration&lt;iOS_GreenRecipes.RecipeCollectionViewCell, iOS_GreenRecipes.RecipeData&gt;"
-                  moduleName = "iOS-GreenRecipes"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/anupamberi/Tutorials/Swift/iOS-GreenRecipes/iOS-GreenRecipes/View%20Controllers/RecipesHomeViewController/RecipesHomeViewController.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "89"
-                  endingLineNumber = "89"
-                  offsetFromSymbolStart = "20">
-               </Location>
-               <Location
-                  uuid = "E7EB5D3F-2D00-429F-8AB0-AAC58835FA1B - e3f87bc8d545aea"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "closure #1 (iOS_GreenRecipes.RecipeCollectionViewCell, Foundation.IndexPath, iOS_GreenRecipes.RecipeData) -&gt; () in iOS_GreenRecipes.RecipesHomeViewController.createRecipeCellRegistration() -&gt; __C.UICollectionView.CellRegistration&lt;iOS_GreenRecipes.RecipeCollectionViewCell, iOS_GreenRecipes.RecipeData&gt;"
-                  moduleName = "iOS-GreenRecipes"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/anupamberi/Tutorials/Swift/iOS-GreenRecipes/iOS-GreenRecipes/View%20Controllers/RecipesHomeViewController/RecipesHomeViewController.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "90"
-                  endingLineNumber = "90"
-                  offsetFromSymbolStart = "358">
-               </Location>
-            </Locations>
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "F0A769F3-413A-4D6F-B011-BAEF6216022F"
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "iOS-GreenRecipes/View Controllers/RecipesHomeViewController/RecipesHomeViewController.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "105"
-            endingLineNumber = "105"
-            landmarkName = "fetchData()"
-            landmarkType = "7">
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "6FD990B7-40DA-4BB6-9BAC-6A0C1D3F0F1F"
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "iOS-GreenRecipes/View Controllers/RecipesHomeViewController/RecipesHomeViewController.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "271"
-            endingLineNumber = "271"
-            landmarkName = "configureDataSource()"
-            landmarkType = "7">
-         </BreakpointContent>
-      </BreakpointProxy>
    </Breakpoints>
 </Bucket>
